visreg: Visualization of Regression Models

Provides a convenient interface for constructing plots to visualize the fit of regression models arising from a wide variety of models in R ('lm', 'glm', 'coxph', 'rlm', 'gam', 'locfit', 'lmer', 'randomForest', etc.)

References

Breheny P and Burchett W. (2017) Visualization of regression models using visreg. *R Journal*, **9**: 56-71. doi:10.32614/RJ-2017-046

See also

Author

Maintainer: Patrick Breheny patrick-breheny@uiowa.edu (ORCID)

Other contributors:

Woodrow Burchett [contributor]

Examples

vignette("quick-start", package="visreg")
#> Warning: vignette ‘quick-start’ not found